# docs/swagger.yaml
basePath: /api
definitions:
  model.Autocomplete:
    properties:
      highlight:
        type: string
      title:
        type: string
    type: object
  model.AutocompleteResp:
    properties:
      autocompletes:
        items:
          $ref: '#/definitions/model.Autocomplete'
        type: array
    type: object
  model.BulkInsertReq:
    properties:
      products:
        items:
          $ref: '#/definitions/model.ProductSearchInsert'
        type: array
    type: object
  model.BulkInsertResp:
    type: object
  model.BulkUpdateReq:
    properties:
      products:
        items:
          $ref: '#/definitions/model.ProductSearchUpdate'
        type: array
    type: object
  model.BulkUpdateResp:
    type: object
  model.ProductSearchCatalogs:
    properties:
      catalog:
        type: string
      count:
        type: integer
    type: object
  model.ProductSearchInsert:
    properties:
      catalog:
        type: string
      cta_url:
        type: string
      description:
        type: string
      image_url:
        type: string
      price:
        type: number
      title:
        type: string
    type: object
  model.ProductSearchResponse:
    properties:
      catalog:
        type: string
      cta_url:
        type: string
      description:
        type: string
      id:
        type: string
      image_url:
        type: string
      price:
        type: number
      title:
        type: string
    type: object
  model.ProductSearchUpdate:
    properties:
      catalog:
        type: string
      cta_url:
        type: string
      description:
        type: string
      id:
        type: string
      image_url:
        type: string
      price:
        type: number
      title:
        type: string
    type: object
  model.SearchCatalogsResp:
    properties:
      catalogs:
        items:
          $ref: '#/definitions/model.ProductSearchCatalogs'
        type: array
    type: object
  model.SearchResponse:
    properties:
      next_cursor:
        type: string
      products:
        items:
          $ref: '#/definitions/model.ProductSearchResponse'
        type: array
    type: object
  model.SearchTotalResp:
    properties:
      total:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a product search API swagger documentation.
  title: Product Search API
paths:
  /products/bulk:
    post:
      consumes:
      - application/json
      description: BulkInsert bulk insert products
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.BulkInsertReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BulkInsertResp'
      summary: BulkInsert bulk insert products
      tags:
      - Products API
  /products/bulk/update:
    post:
      consumes:
      - application/json
      description: BulkUpdate bulk update products
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.BulkUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BulkUpdateResp'
      summary: BulkUpdate bulk update products
      tags:
      - Products API
  /search:
    get:
      consumes:
      - application/json
      description: Search product from given q
      parameters:
      - description: q
        in: query
        name: q
        type: string
      - description: catalog
        in: query
        name: catalog
        type: string
      - description: size
        in: query
        name: size
        type: integer
      - description: sort_by
        in: query
        name: sort_by
        type: string
      - description: next_cursor
        in: query
        name: next_cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SearchResponse'
      summary: Search product from given q
      tags:
      - Search API
  /search/autocomplete:
    get:
      consumes:
      - application/json
      description: Search autocomplete from given q
      parameters:
      - description: q
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.AutocompleteResp'
      summary: Search autocomplete from given q
      tags:
      - Search API
  /search/catalogs:
    get:
      consumes:
      - application/json
      description: Search product catalogs from given q
      parameters:
      - description: q
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SearchCatalogsResp'
      summary: Search product catalogs from given q
      tags:
      - Search API
  /search/total:
    get:
      consumes:
      - application/json
      description: Search product total from given q
      parameters:
      - description: q
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SearchTotalResp'
      summary: Search product total from given q
      tags:
      - Search API
swagger: "2.0"